standardisation of the cancellation-sf-case module (no longer reads identity cookies or handles test-users - that has shifted to the consumer)
