+++ handlers/cancellation-sf-cases-api/README.md
@@ -1,0 +1,14 @@
+# cancellation-sf-cases-api
+API Gateway fronted lambda function for creating (`POST`) and updating (`PATCH`) SalesForce 'Cases', called during cancellation flow(s) within 'manage-frontend'.
+
+All endpoints require...
+
+- `x-api-key` header (generated by the `CancellationSFCasesApiKey` in the CloudFormation)
+- and either...
+  - `x-identity-id` header which specifies the identityID of the user to request data for _(sent in the `manage-frontend` use-case)_
+  
+#### Handling Multiple Environments
+See [handlers/README](../README.md#3rd-party-service-environments) for information on how this aspect works.
+
+### TODO describe the endpoints in detail
+
