+++ handlers/delivery-records-api/README.md
@@ -2,16 +2,17 @@
 
 This api provides access to delivery records for a subscription. Delivery records are stored in Salesforce. 
 
-### Usage
-
 All endpoints require...
 
 - `x-api-key` header (generated by the `DeliveryRecordsApiKey` in the CloudFormation)
 - and either...
   - `x-identity-id` header which specifies the identityID of the user to request data for _(sent in the `manage-frontend` use-case)_
   - `x-salesforce-contact-id` header which specifies the Salesforce contact ID of the user to request data for _(sent in the CSR UI (in Salesforce) use-case)_
+  
+#### Handling Multiple Environments
+See [handlers/README](../README.md#3rd-party-service-environments) for information on how this aspect works.
 
-#### **`GET`** `/{STAGE}/delivery-records/{SUBSCRIPTION_NAME}?startDate={yyyy-MM-dd}&endDate={yyyy-MM-dd}`
+### **`GET`** `/{STAGE}/delivery-records/{SUBSCRIPTION_NAME}?startDate={yyyy-MM-dd}&endDate={yyyy-MM-dd}`
 
 - Returns the delivery records for a subscription, optionally filtered based on delivery date.
 - Returns 404 if the user doesn't have access to the subscription.
@@ -80,7 +81,7 @@ All endpoints require...
 }
 ```
 
-#### `POST` `/{STAGE}/delivery-records/{SUBSCRIPTION_NAME}`
+### `POST` `/{STAGE}/delivery-records/{SUBSCRIPTION_NAME}`
 
 - Creates a delivery problem case, linking the provided delivery records to the new case, then **returns the same as the `GET`** once the changes have been made.
 - Optionally takes a `repeatDeliveryProblem` top-level field to mark the case as a 'repeat delivery problem'.
